fix(pr): extract decorate backend url function
