+++ src/lib/elastic.js
@@ -3,6 +3,7 @@ const _ = require('lodash')
 const fs = require('fs');
 const jsonFile = require('jsonfile')
 const es = require('@elastic/elasticsearch')
+const querystring = require('querystring')
 
 function _updateQueryStringParameter (uri, key, value) {
   var re = new RegExp('([?&])' + key + '=.*?(&|#|$)', 'i');
@@ -29,6 +30,39 @@ function adjustIndexName (indexName, entityType, config) {
   } else {
     return `${indexName}_${entityType}`
   }
+}
+
+function decorateBackendUrl (entityType, url, req, config) {
+  if (config.elasticsearch.useRequestFilter && typeof config.entities[entityType] === 'object') {
+    const urlParts = url.split('?')
+    const { includeFields, excludeFields } = config.entities[entityType]
+
+    const filteredParams = Object.keys(req.query)
+      .filter(key => !config.elasticsearch.requestParamsBlacklist.includes(key))
+      .reduce((object, key) => {
+        object[key] = req.query[key]
+        return object
+      }, {})
+
+    let _source_include = includeFields || []
+    let _source_exclude = excludeFields || []
+
+    if (!config.elasticsearch.overwriteRequestSourceParams) {
+      const requestSourceInclude = req.query._source_include || []
+      const requestSourceExclude = req.query._source_exclude || []
+      _source_include = [...includeFields, ...requestSourceInclude]
+      _source_exclude = [...excludeFields, ...requestSourceExclude]
+    }
+
+    const urlParams = {
+      ...filteredParams,
+      _source_include,
+      _source_exclude
+    }
+    url = `${urlParts[0]}?${querystring.stringify(urlParams)}`
+  }
+
+  return url
 }
 
 function adjustBackendProxyUrl (req, indexName, entityType, config) {
@@ -58,7 +92,8 @@ function adjustBackendProxyUrl (req, indexName, entityType, config) {
   if (!url.startsWith('http')) {
     url = config.elasticsearch.protocol + '://' + url
   }
-  return url
+
+  return decorateBackendUrl(entityType, url, req, config)
 }
 
 function adjustQuery (esQuery, entityType, config) {
